==> eicu_processing/2_diagnosis trim.ipynb <==
import numpy as np
import pandas as pd
import psycopg2
import getpass
# for configuring connection 
from configobj import ConfigObj
import os
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from utils import *

# Create a database connection using settings from config file
config='../db/config.ini'

# connection info
conn_info = dict()
if os.path.isfile(config):
    config = ConfigObj(config)
    conn_info["sqluser"] = config['username']
    conn_info["sqlpass"] = config['password']
    conn_info["sqlhost"] = config['host']
    conn_info["sqlport"] = config['port']
    conn_info["dbname"] = config['dbname']
    conn_info["schema_name"] = config['schema_name']
else:
    conn_info["sqluser"] = 'postgres'
    conn_info["sqlpass"] = ''
    conn_info["sqlhost"] = '192.168.60.144'
    conn_info["sqlport"] = 6432
    conn_info["dbname"] = 'eicu'
    conn_info["schema_name"] = 'public,eicu_crd'
    
# Connect to the eICU database
print('Database: {}'.format(conn_info['dbname']))
print('Username: {}'.format(conn_info["sqluser"]))
if conn_info["sqlpass"] == '':
    # try connecting without password, i.e. peer or OS authentication
    try:
        if (conn_info["sqlhost"] == '192.168.60.144') & (conn_info["sqlport"]=='6432'):
            con = psycopg2.connect(dbname=conn_info["dbname"],
                                   user=conn_info["sqluser"])            
        else:
            con = psycopg2.connect(dbname=conn_info["dbname"],
                                   host=conn_info["sqlhost"],
                                   port=conn_info["sqlport"],
                                   user=conn_info["sqluser"])
    except:
        conn_info["sqlpass"] = getpass.getpass('Password: ')

        con = psycopg2.connect(dbname=conn_info["dbname"],
                               host=conn_info["sqlhost"],
                               port=conn_info["sqlport"],
                               user=conn_info["sqluser"],
                               password=conn_info["sqlpass"])
query_schema = 'set search_path to ' + conn_info['schema_name'] + ';'

from sqlalchemy import create_engine,text

con = create_engine(
    f'postgresql://{conn_info["sqluser"]}:{conn_info["sqlpass"]}@{conn_info["sqlhost"]}:{conn_info["sqlport"]}/{conn_info["dbname"]}',
    connect_args={'options': '-c search_path=eicu_crd'}
)

create_table_query = query_schema + """
drop materialized view if exists diagnoses cascade;
create materialized view diagnoses as
  -- for past medical history:
select l.patient,l.apacheadmissiondx, ph.pasthistorypath as diagnosisstring
from pasthistory as ph
inner join labels as l on l.patient = ph.patientunitstayid
where ph.pasthistoryoffset > 0
      and ph.pasthistorypath LIKE '%Organ Systems%'      
"""
with con.begin() as connection:
    connection.execute(text(create_table_query)) 
    
select_query = "SELECT * FROM diagnoses;"
df_dx = pd.read_sql_query(select_query, con)
print("there are {} patiets and {} records".format(df_dx["patient"].nunique(),df_dx["patient"].count()))

## process the diagnosis 

flat =df_dx[['patient','apacheadmissiondx']].drop_duplicates()
diagnoses = df_dx.copy()

diagnoses 

query = query_schema + """
select distinct(patient)
from id
order by patient;
"""
id = pd.read_sql_query(query, con)

common_patients = set( diagnoses['patient'])& set(id['patient'])
diagnoses = diagnoses[diagnoses['patient'].isin(common_patients)]
print("there are {} patiets and {} records".format(diagnoses["patient"].nunique(),diagnoses["patient"].count()))

diagnoses=diagnoses.drop(columns=['apacheadmissiondx'])
diagnoses

df1 = diagnoses.copy()
df1['diagnosisstring'] = df1['diagnosisstring'].str.replace('notes/Progress Notes/Past History/Organ Systems/', ' ')
df1['diagnosisstring'] = df1['diagnosisstring'].str.replace('|', '/')
split = df1['diagnosisstring'].str.split('/')
df1['first'] = split.str[0]
df1['second'] = split.str[1]
df1['third'] = split.str[2]
df1['fourth'] = split.str[3]
df1

frequency4 = pd.DataFrame()
frequency4['noAnnotations'] = df1['fourth'].value_counts() 
frequency4['%'] = df1['fourth'].value_counts() / len(df1.index) * 100
frequency4

frequency1 = pd.DataFrame()
frequency1['noAnnotations'] = df1['first'].value_counts() 
frequency1['%'] = df1['first'].value_counts() / len(df1.index) * 100
frequency1

frequency2 = pd.DataFrame()
frequency2['noAnnotations'] = df1['second'].value_counts() 
frequency2['%'] = df1['second'].value_counts() / len(df1.index) * 100
frequency2

len(frequency2)

from bigtree import Node, list_to_tree, tree_to_dot, dataframe_to_tree, tree_to_pillow

for disease in frequency1.index.get_level_values('first'):
    df_ = df1[df1['first'] == disease]
    df_.set_index(['patient'], inplace = True)
    
    paths_mult_ = df_.groupby('diagnosisstring', dropna = True)['diagnosisstring'].size().to_frame('noAnnotations')
    paths_mult_.reset_index(inplace=True)
    paths_mult_.sort_values('noAnnotations', ascending = False, inplace = True)

    paths_ = list(pd.unique(df_['diagnosisstring']))
    root_ = dataframe_to_tree(paths_mult_, sep = '/')
    root_.show(attr_list=["noAnnotations"])
    pillow_image_ = tree_to_pillow(root_)

df1['first'] = df1['first'].str.strip()
from tabulate import tabulate

diseases = ['Cardiovascular (R)', 'Endocrine (R)', 'Pulmonary', 'Neurologic', 'Hematology', 'Renal  (R)', 'Gastrointestinal (R)', 'Rheumatic', 'Infectious Disease (R)']

for disease in diseases:
 
    df_disease = df1[df1['first'] == disease].copy()

    freq = pd.DataFrame()
    freq['noAnnotations'] = df_disease['second'].value_counts()
    freq['%'] = round(df_disease['second'].value_counts() / len(df_disease.index) * 100, 2)

    table = tabulate(freq, headers="keys", tablefmt="psql", showindex=True)
    print(f"The frequency of {disease}:")
    print(table)
    print("------------------------------")

df1['diagnosisstring'] = df1['diagnosisstring'].str.split('/').str[:2].str.join('/')
df1 = df1.drop(columns=['third','fourth'])
df1

csv= '/home/mei/nas/docker/thesis/data/csv/'
df1.to_csv(csv+'diagnoses_trim_level.csv')

hdf= '/home/mei/nas/docker/thesis/data/hdf/'
df1.to_hdf(hdf+'final_diagnoses_trim_level.h5', key='df')

df1

hdf= '/home/mei/nas/docker/thesis/data/hdf/'
df1 = pd.read_hdf(hdf+'final_diagnoses_trim_level.h5', key='df')

df1 

frequency1 = pd.DataFrame()
frequency1['noAnnotations'] = df1['first'].value_counts() 
frequency1['%'] = df1['first'].value_counts() / len(df1.index) * 100
frequency1

from bigtree import Node, list_to_tree, tree_to_dot, dataframe_to_tree, tree_to_pillow

for disease in frequency1.index.get_level_values('first'):
    df_ = df1[df1['first'] == disease]
    df_.set_index(['patient'], inplace = True)
    
    paths_mult_ = df_.groupby('diagnosisstring', dropna = True)['diagnosisstring'].size().to_frame('noAnnotations')
    paths_mult_.reset_index(inplace=True)
    paths_mult_.sort_values('noAnnotations', ascending = False, inplace = True)

    paths_ = list(pd.unique(df_['diagnosisstring']))
    root_ = dataframe_to_tree(paths_mult_, sep = '/')
    root_.show(attr_list=["noAnnotations"])
    pillow_image_ = tree_to_pillow(root_)
    
    ## thesis: 解释 为什么要用graph 

diagnoses = df1.copy()
# diagnoses ['diagnosisstring'] =diagnoses['diagnosisstring'].str.replace('/', '|')

unique_diagnoses = diagnoses['diagnosisstring'].unique()
codes_dict, mapping_dict, count, words_dict = build_mapping_dict(unique_diagnoses)

# 1) get unique patient_id
patients = diagnoses['patient'].unique()
n_patients = len(patients)        # 3072

# 2) build sparse matrix to store the diagnosis codes  shape = (n_patients, count)
sparse_mat = np.zeros((n_patients, count), dtype=int)
patient_to_idx = {pid: i for i, pid in enumerate(patients)}

# 3) visit each patient, and get the diagnosis codes
for pid, grp in diagnoses.groupby('patient'):
    codes = []
    for diag in grp['diagnosisstring']:
        codes += mapping_dict.get(diag, [])
    codes = [c for c in codes if 0 <= c < count]
    i = patient_to_idx[pid]
    sparse_mat[i, codes] = 1

# 4) convert to DataFrame  ,shape = (n_patients, count)
sparse_df = pd.DataFrame(sparse_mat, index=patients, columns=range(count))

# 5)
print("Before drop:", sparse_df.shape)

# 6) calculate the cutoff count
cutoff_cnt = max(1, int(0.01 * n_patients)) 

# 7)drop the codes that are not used by at least cutoff_cnt patients
drop_once = find_unnecessary_codes(codes_dict)
drop_rare = find_rare_codes(cutoff_cnt, sparse_df)
all_drop = set(drop_once + drop_rare)
sparse_df = sparse_df.drop(columns=all_drop, errors='ignore')
sparse_df.rename(columns=words_dict, inplace=True)

print(f"Dropped {len(drop_once)} useless + {len(drop_rare)} rare codes")
print("After  drop:", sparse_df.shape)

sparse_df = sparse_df.reset_index().rename(columns={'index': 'patient'})
sparse_df

sparse_df.columns

words_dict 

hdf= '/home/mei/nas/docker/thesis/data/hdf/'
sparse_df.to_hdf(hdf+'final_diagnoses_trim.h5', key='df')
