==> eicu_processing/5_lab.ipynb <==
import numpy as np
import pandas as pd
import itertools
import gc
import psycopg2
import getpass
# for configuring connection 
from configobj import ConfigObj
import os
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from utils import *

# Create a database connection using settings from config file
config='../db/config.ini'

# connection info
conn_info = dict()
if os.path.isfile(config):
    config = ConfigObj(config)
    conn_info["sqluser"] = config['username']
    conn_info["sqlpass"] = config['password']
    conn_info["sqlhost"] = config['host']
    conn_info["sqlport"] = config['port']
    conn_info["dbname"] = config['dbname']
    conn_info["schema_name"] = config['schema_name']
else:
    conn_info["sqluser"] = 'postgres'
    conn_info["sqlpass"] = ''
    conn_info["sqlhost"] = '192.168.60.144'
    conn_info["sqlport"] = 6432
    conn_info["dbname"] = 'eicu'
    conn_info["schema_name"] = 'public,eicu_crd'
    
# Connect to the eICU database
print('Database: {}'.format(conn_info['dbname']))
print('Username: {}'.format(conn_info["sqluser"]))
if conn_info["sqlpass"] == '':
    # try connecting without password, i.e. peer or OS authentication
    try:
        if (conn_info["sqlhost"] == '192.168.60.144') & (conn_info["sqlport"]=='6432'):
            con = psycopg2.connect(dbname=conn_info["dbname"],
                                   user=conn_info["sqluser"])            
        else:
            con = psycopg2.connect(dbname=conn_info["dbname"],
                                   host=conn_info["sqlhost"],
                                   port=conn_info["sqlport"],
                                   user=conn_info["sqluser"])
    except:
        conn_info["sqlpass"] = getpass.getpass('Password: ')

        con = psycopg2.connect(dbname=conn_info["dbname"],
                               host=conn_info["sqlhost"],
                               port=conn_info["sqlport"],
                               user=conn_info["sqluser"],
                               password=conn_info["sqlpass"])
query_schema = 'set search_path to ' + conn_info['schema_name'] + ';'

from sqlalchemy import create_engine,text

con = create_engine(
    f'postgresql://{conn_info["sqluser"]}:{conn_info["sqlpass"]}@{conn_info["sqlhost"]}:{conn_info["sqlport"]}/{conn_info["dbname"]}',
    connect_args={'options': '-c search_path=eicu_crd'}
)

create_table_query =  query_schema +"""
drop materialized view if exists timeserieslab cascade;
create materialized view timeserieslab as
  select l.patientunitstayid as patient, l.labresultoffset, l.labname,l.labresult
    from lab as l
    inner join id 
      on id.patient = l.patientunitstayid -- only extract data for the cohort
    where l.labresultoffset > 0;

"""

with con.begin() as connection:
    connection.execute(text(create_table_query)) 

select_query = "SELECT * FROM timeserieslab;"
df_lab = pd.read_sql_query(select_query, con)

print("there are {} patiets and {} records in lab test table".format(df_lab["patient"].nunique(),df_lab["patient"].count()))

timeseries_lab = df_lab.copy()
timeseries_lab = timeseries_lab[timeseries_lab['labresultoffset'] <= 11*24*60]
timeseries_lab.set_index(['patient','labresultoffset'], inplace=True)
timeseries_lab.rename(round_up, level = 'labresultoffset', inplace = True)
timeseries_lab.sort_index(inplace=True)

lab = process_vital_signs(timeseries_lab,'labresultoffset',lab=True)

hdf= '/home/mei/nas/docker/thesis/data/hdf/'
vital = pd.read_hdf(hdf + 'vital_6_8.h5', key='vital_6_8')
l = lab.copy()
l.columns = l.columns.droplevel(0)
## get common patients
common_patients = list(set(l.index.get_level_values(0)) & set(vital.index.get_level_values(0)))
print("there are {} common patients".format(len(common_patients)))

l = l.loc[l.index.get_level_values(0).isin(common_patients)]
vital = vital.loc[vital.index.get_level_values(0).isin(common_patients)]
print("there are {} patients in vital and {} patients in lab".format(vital.index.get_level_values(0).nunique(),l.index.get_level_values(0).nunique()))
print("there are {} records in vital and {} records in lab".format(vital.shape[0],l.shape[0]))
print('==> Combining data together...')
merged =  pd.concat([l, vital], axis=0, sort=False)
print("there are {} patients in the merged table and {} records.".format(len(list(merged.index.get_level_values(0).unique())),len(merged)))

patients = merged.index.unique(level=0)
gen_chunks = gen_patient_chunk(patients, merged)
header = True
print('==> Initiating main processing loop...')

for i, patient_chunk in enumerate(gen_chunks, start=1):
    final = resample(patient_chunk)
    final.to_hdf(hdf + "final_timeseries.h5", key="df", mode="a", complevel=5, complib="zlib", format="table", append=True)

    print(f'==> Processed {i * 500} patients...')
    
    del patient_chunk # free up memory
    gc.collect()

hdf= '/home/mei/nas/docker/thesis/data/hdf/'
final_ts= pd.read_hdf(hdf+ "final_timeseries.h5", key="df") # key is the name of the key in the hdf file
final_ts

##  the max time step
print("the max time step is {}".format(final_ts.index.get_level_values(1).max()))

# 1. 每个病人的最大时间长度
patient_max_time = final_ts.reset_index().groupby('patient')['time'].max()

# 2. 按500分组统计
step = 500
max_time = patient_max_time.max()
bins = np.arange(0, max_time + step, step)

# 3. 用pd.cut分箱
bin_labels = [f"{int(bins[i])}-{int(bins[i+1])}" for i in range(len(bins)-1)]
patient_bins = pd.cut(patient_max_time, bins=bins, labels=bin_labels, right=False)

# 4. 每个区间的病人数
counts = patient_bins.value_counts().sort_index()

for label, num in counts.items():
    print(f"Max time {label}: {num} patients")

final_ts.columns
# save the feature names
features = final_ts.columns
features.to_series().to_csv('/home/mei/nas/docker/thesis/data/hdf/features.txt', index=False, header=False)

with open('/home/mei/nas/docker/thesis/data/hdf/features.txt', 'r') as f:
    features = [line.strip() for line in f]
features

ts =final_ts.copy()
ts = ts.reset_index()
ts = ts[['patient','time']]
ts

with con.begin() as conn:
    conn.execute(text("DROP MATERIALIZED VIEW IF EXISTS id CASCADE;"))
    conn.execute(text("DROP TABLE IF EXISTS temp_id CASCADE;"))

ts.to_sql("temp_id", con, if_exists="replace", index=True, method="multi")

with con.begin() as connection:
    connection.execute(text("DROP TABLE IF EXISTS id CASCADE;"))
    connection.execute(text("""
                            
        CREATE MATERIALIZED VIEW id AS
        SELECT patient, time
        FROM temp_id;
        
    """))
    

select_query = "SELECT patient FROM id;"
df_id = pd.read_sql_query(select_query, con)
print("there are {} patiets and {} records".format(df_id["patient"].nunique(),df_id["patient"].count()))
