==> eicu_processing/7_split_train_test.ipynb <==
import pandas as pd
import os
from utils import *

def split_train_test(eICU_path, flat_features, diagnoses,
                     timeseries,timeseries_15,timeseries_30,timeseries_60,
                     risks     ,risks_15,     risks_30,     risks_60,
                     seed=9):

    # Split data into train, validation, and test sets
    train, test = train_test_split(flat_features.index, test_size=0.15, random_state=seed)
    train, val = train_test_split(train, test_size=0.15/0.85, random_state=seed)

    print('==> Loading data for splitting...')
    # Load datasets


    # Process and save partitions
    for partition_name, partition in zip(['train', 'val', 'test'], [train, val, test]):
        print(f'==> Preparing {partition_name} data...')
        folder_path = create_folder(eICU_path, partition_name)
        stays = shuffle_stays(partition, seed=seed)
        stays_path = os.path.join(folder_path, 'stays.txt')
         
        with open(stays_path, 'w') as f:
            for stay in stays:
                f.write(f"{stay}\n")
        for table_name, table in zip(['flat', 'diagnoses', 'timeseries','timeseries_15','timeseries_30','timeseries_60',
                                      'risks','risks_15','risks_30','risks_60'], 
                                     [ flat_features, diagnoses,
                                       timeseries,timeseries_15,timeseries_30,timeseries_60,
                                       risks,risks_15,risks_30,risks_60
                                       ]):
            
            process_table(table_name, table, stays, folder_path)
        print(f'==> {partition_name} data saved!\n')
        
    print("\n==== Dataset Sizes ====")
    for partition_name, partition in zip(['train', 'val', 'test'], [train, val, test]):
        print(f"**{partition_name} set:**")
        print(f"- Flat Features: {flat_features.loc[flat_features.index.isin(partition)].shape}")
        print(f"- Diagnoses: {diagnoses.loc[diagnoses.index.isin(partition)].shape}")
        
        print(f"- Time Series: {timeseries.loc[timeseries.index.isin(partition)].shape}")
        print(f"- Labels: {risks.loc[risks.index.isin(partition)].shape}\n")
        
        print(f"- Resample 15min Time Series: {timeseries_15.loc[timeseries_15.index.isin(partition)].shape}")
        print(f"- Resample 15min Labels: {risks_15.loc[risks_15.index.isin(partition)].shape}\n")
        
        print(f"- Resample 30min Time Series: {timeseries_30.loc[timeseries_30.index.isin(partition)].shape}")
        print(f"- Resample 30min Labels: {risks_30.loc[risks_30.index.isin(partition)].shape}\n")
        
        print(f"- Resample 60min Time Series: {timeseries_60.loc[timeseries_60.index.isin(partition)].shape}")
        print(f"- Resample 60min Labels: {risks_60.loc[risks_60.index.isin(partition)].shape}\n")

  
    print('==> Splitting complete!')
    return

def save_to_h5py(data_path, partition, table_name):
    import os, h5py, pandas as pd
    partition_path = os.path.join(data_path, partition)
    
    input_file = os.path.join(partition_path, f"{table_name}.h5")
    data = pd.read_hdf(input_file)
    
    output_file = os.path.join(partition_path, f"{table_name}_each_patient.h5")
    with h5py.File(output_file, 'w') as h5f:
        for patient_id, group in data.groupby('patient'):
            h5f.create_dataset(str(patient_id), data=group.values)
    print(f"Saved h5py for {partition}/{table_name}")

hdf= '/home/mei/nas/docker/thesis/data/hdf/'

flat_features = pd.read_hdf(os.path.join(hdf, 'final_flat_drug.h5')).set_index('patient')
diagnoses = pd.read_hdf(os.path.join(hdf, 'final_diagnoses_trim_level.h5')).set_index('patient')
timeseries = pd.read_hdf(os.path.join(hdf, 'final_timeseries.h5')).reset_index().set_index('patient')
risks = pd.read_hdf(os.path.join(hdf, 'final_risk_scores.h5')).set_index('patient')

timeseries_15 = pd.read_hdf(os.path.join(hdf, 'final_timeseries_15min.h5')).reset_index().set_index('patient')
risks_15 = pd.read_hdf(os.path.join(hdf, 'final_risk_scores_15min.h5'))
timeseries_30 = pd.read_hdf(os.path.join(hdf, 'final_timeseries_30min.h5')).reset_index().set_index('patient')
risks_30 = pd.read_hdf(os.path.join(hdf, 'final_risk_scores_30min.h5'))
timeseries_60 = pd.read_hdf(os.path.join(hdf, 'final_timeseries_60min.h5')).reset_index().set_index('patient')
risks_60 = pd.read_hdf(os.path.join(hdf, 'final_risk_scores_60min.h5'))


split_train_test(hdf,flat_features, diagnoses,
                timeseries,timeseries_15,timeseries_30,timeseries_60,
                risks,risks_15,risks_30,risks_60
                 )

hdf= '/home/mei/nas/docker/thesis/data/hdf/'

TABLES = [
    'timeseries', 'timeseries_15', 'timeseries_30', 'timeseries_60',
    'risks',      'risks_15',      'risks_30',      'risks_60'
]
for partition in ['train', 'val', 'test']:
    for tbl in TABLES:
        save_to_h5py(hdf, partition, tbl)

risks_60

timeseries

hdf= '/home/mei/nas/docker/thesis/data/hdf/'
flat_features = pd.read_hdf(os.path.join(hdf, 'final_flat_drug.h5')).set_index('patient')
flat_features

flat_features.columns
# save the feature names
features = flat_features.columns
features.to_series().to_csv('/home/mei/nas/docker/thesis/data/hdf/flat_features.txt', index=False, header=False)

with open('/home/mei/nas/docker/thesis/data/hdf/flat_features.txt', 'r') as f:
    features = [line.strip() for line in f]
features

diagnoses = pd.read_hdf(os.path.join(hdf, 'final_diagnoses_trim_level.h5')).set_index('patient')
diagnoses

feature_names = diagnoses['first'] + '/' + diagnoses['second']
feature_names = feature_names.unique().tolist()

feature_names = pd.Series(feature_names)
feature_names.to_csv('/home/mei/nas/docker/thesis/data/hdf/diagnosis_features.txt', index=False, header=False)

with open('/home/mei/nas/docker/thesis/data/hdf/diagnosis_features.txt', 'r') as f:
    features = [line.strip() for line in f]
features

hdf= '/home/mei/nas/docker/thesis/data/hdf/'

risk_test = pd.read_hdf(os.path.join(hdf, 'test/risks_60.h5'))
risk_test

flat_test = pd.read_hdf(os.path.join(hdf, 'test/flat.h5'))
flat_test

diagnoses_test = pd.read_hdf(os.path.join(hdf, 'test/diagnoses.h5'))
diagnoses_test

import h5py
hdf= '/home/mei/nas/docker/thesis/data/hdf/'
risk_train  = h5py.File(os.path.join(hdf, 'train/risks_each_patient.h5'), 'r')
risk_val = h5py.File(os.path.join(hdf, 'val/risks_each_patient.h5'), 'r')
risk_test = h5py.File(os.path.join(hdf, 'test/risks_each_patient.h5'), 'r')


risk_train_15 = h5py.File(os.path.join(hdf, 'train/risks_15_each_patient.h5'), 'r')
risk_val_15 = h5py.File(os.path.join(hdf, 'val/risks_15_each_patient.h5'), 'r')
risk_test_15 = h5py.File(os.path.join(hdf, 'test/risks_15_each_patient.h5'), 'r')

risk_train_30 = h5py.File(os.path.join(hdf, 'train/risks_30_each_patient.h5'), 'r')
risk_val_30 = h5py.File(os.path.join(hdf, 'val/risks_30_each_patient.h5'), 'r')
risk_test_30 = h5py.File(os.path.join(hdf, 'test/risks_30_each_patient.h5'), 'r')

risk_train_60 = h5py.File(os.path.join(hdf, 'train/risks_60_each_patient.h5'), 'r')
risk_val_60 = h5py.File(os.path.join(hdf, 'val/risks_60_each_patient.h5'), 'r')
risk_test_60 = h5py.File(os.path.join(hdf, 'test/risks_60_each_patient.h5'), 'r')

import pandas as pd

def read_h5_file_to_df(h5_file, columns):
    all_data = []
    for pid in h5_file.keys():
        data = h5_file[pid][:]
        df = pd.DataFrame(data, columns=columns)
        df["patient"] = int(pid)
        all_data.append(df)
    return pd.concat(all_data, ignore_index=True)

def summarize_by_risk(df, name=""):
    grouped = df.groupby("discharge_risk_category")
    summary = grouped.agg(
        patient_count=("patient", lambda x: x.nunique()),
        record_count=("patient", "count")
    ).sort_index()

    print(f"\ Summary for {name.upper()}:")
    print(summary)
    return summary

COLUMNS = ['time', 'gender', 'age', 'dischargeweight', 'unitdischargestatus',
           'discharge_risk_category', 'actualiculos','time_to_discharge', 'risk_score']

df_train = read_h5_file_to_df(risk_train, COLUMNS)
df_val = read_h5_file_to_df(risk_val, COLUMNS)
df_test = read_h5_file_to_df(risk_test, COLUMNS)

summarize_by_risk(df_train, "train")
summarize_by_risk(df_val, "val")
summarize_by_risk(df_test, "test")

train_15 = read_h5_file_to_df(risk_train_15, COLUMNS)
val_15 = read_h5_file_to_df(risk_val_15, COLUMNS)
test_15 = read_h5_file_to_df(risk_test_15, COLUMNS)
summarize_by_risk(train_15, "train 15min")
summarize_by_risk(val_15, "val 15min")
summarize_by_risk(test_15, "test 15min")

train_30= read_h5_file_to_df(risk_train_30, COLUMNS)
val_30 = read_h5_file_to_df(risk_val_30, COLUMNS)
test_30 = read_h5_file_to_df(risk_test_30, COLUMNS)
summarize_by_risk(train_30, "train 30min")
summarize_by_risk(val_30, "val 30min")
summarize_by_risk(test_30, "test 30min")


train_60 = read_h5_file_to_df(risk_train_60, COLUMNS)
val_60 = read_h5_file_to_df(risk_val_60, COLUMNS)
test_60 = read_h5_file_to_df(risk_test_60, COLUMNS)
summarize_by_risk(train_60, "train 60min")
summarize_by_risk(val_60, "val 60min")
summarize_by_risk(test_60, "test 60min")

import h5py
hdf= '/home/mei/nas/docker/thesis/data/hdf/'

risk_train_60 = h5py.File(os.path.join(hdf, 'train/risks_60_each_patient.h5'), 'r')
risk_val_60 = h5py.File(os.path.join(hdf, 'val/risks_60_each_patient.h5'), 'r')
risk_test_60 = h5py.File(os.path.join(hdf, 'test/risks_60_each_patient.h5'), 'r')

def read_h5_file_to_df(h5_file, columns):
    all_data = []
    for pid in h5_file.keys():
        data = h5_file[pid][:]
        df = pd.DataFrame(data, columns=columns)
        df["patient"] = int(pid)
        all_data.append(df)
    return pd.concat(all_data, ignore_index=True)

def summarize_by_mortality(df, name=""):
    grouped = df.groupby("unitdischargestatus")
    summary = grouped.agg(
        patient_count=("patient", lambda x: x.nunique()),
        record_count=("patient", "count")
    ).sort_index()

    print(f"\ Summary for {name.upper()}:")
    print(summary)
    return summary

COLUMNS = ['time', 'gender', 'age', 'dischargeweight', 'unitdischargestatus',
           'discharge_risk_category', 'actualiculos','time_to_discharge', 'risk_score']


df_test = read_h5_file_to_df(risk_test_60, COLUMNS)
df_train = read_h5_file_to_df(risk_train_60, COLUMNS)
df_val = read_h5_file_to_df(risk_val_60, COLUMNS)
summarize_by_mortality(df_train, "train 60min mortality")
summarize_by_mortality(df_val, "val 60min mortality")
summarize_by_mortality(df_test, "test 60min mortality")
